Renamed some icon files to void ambiguity with similarly named partials Added keyboard handler for paragraph menu Passed legacyMode prop
