+++ library/src/scripts/components/icons/editorIcons.tsx
