+++ plugins/rich-editor/src/scripts/components/toolbars/pieces/MenuItem.test.tsx
@@ -8,7 +8,7 @@ import React from "react";
 import { shallow } from "enzyme";
 import sinon from "sinon";
 import MenuItem from "@rich-editor/components/toolbars/pieces/MenuItem";
-import { bold } from "@library/components/icons/editor";
+import { bold } from "@library/components/icons/editorIcons";
 
 const noop = () => {
     return;
